Code review and fix brittleness issue in stdlib
