+++ Source/DafnyCore/AST/Attributes.cs
@@ -66,7 +66,7 @@ public class Attributes : NodeWithComputedRange, ICanFormat {
 
   public readonly Attributes? Prev;
 
-  [ParseConstructor]
+  [ParsedConstructor]
   public Attributes(IOrigin origin, string name, List<Expression> args, Attributes? prev) : base(origin) {
     Name = name;
     Args = args;
