+++ Source/DafnyCore/AST/Expressions/Specification.cs
@@ -17,7 +17,7 @@ public class Specification<T> : NodeWithComputedRange, IAttributeBearingDeclarat
     Attributes = null;
   }
 
-  [ParseConstructor]
+  [ParsedConstructor]
   public Specification(IOrigin origin, List<T> expressions, Attributes attributes) : base(origin) {
     Expressions = expressions;
     Attributes = attributes;
