+++ Source/DafnyCore/AST/Expressions/Variables/Formal.cs
@@ -22,7 +22,7 @@ public class Formal : NonglobalVariable {
       isOld, isNameOnly, isOlder, nameForCompilation) {
   }
 
-  [ParseConstructor]
+  [ParsedConstructor]
   public Formal(IOrigin origin, Name nameNode, Type type, bool inParam, bool isGhost, Expression defaultValue,
     Attributes attributes = null,
     bool isOld = false, bool isNameOnly = false, bool isOlder = false, string nameForCompilation = null)
