+++ Source/DafnyCore/AST/Members/MethodOrFunction.cs
@@ -29,7 +29,7 @@ public abstract class MethodOrFunction : MemberDecl, ICodeContainer {
   public readonly IOrigin? SignatureEllipsis;
   public override bool IsRefining => SignatureIsOmitted;
 
-  [ParseConstructor]
+  [ParsedConstructor]
   protected MethodOrFunction(IOrigin origin, Name nameNode, bool hasStaticKeyword, bool isGhost,
     Attributes attributes, IOrigin signatureEllipsis, List<TypeParameter> typeArgs, List<Formal> ins,
     List<AttributedExpression> req,
