+++ Source/DafnyCore/AST/SourceOrigin.cs
@@ -16,7 +16,7 @@ public class SourceOrigin : IOrigin, IComparable<SourceOrigin> {
   public bool InclusiveEnd => endToken != null;
   public bool IncludesRange => true;
 
-  [ParseConstructor]
+  [ParsedConstructor]
   public SourceOrigin(Token startToken, Token? endToken, Token? center = null) {
     this.endToken = endToken;
     StartToken = startToken;
