+++ Source/DafnyStandardLibraries/src/Std/JSON/ZeroCopy/Serializer.dfy
@@ -278,6 +278,8 @@ module Std.JSON.ZeroCopy.Serializer {
   function StructuralView(st: Structural<View>, writer: Writer) : (wr: Writer)
     ensures wr.Bytes() == writer.Bytes() + Spec.Structural(st, Spec.View)
   {
+    hide *;
+    reveal Writer.Append, Spec.Structural, Spec.View;
     writer.Append(st.before).Append(st.t).Append(st.after)
   }
