+++ Source/Scripts/Scripts.csproj
@@ -8,17 +8,15 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="System.CommandLine" Version="2.0.0-beta4.22272.1" />
-        <PackageReference Include="NJsonSchema" Version="11.1.0" />
+        <PackageReference Include="System.CommandLine" Version="2.0.0-beta4.22272.1" />
         <PackageReference Include="Microsoft.CodeAnalysis" Version="4.12.0" />
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\DafnyCore\DafnyCore.csproj" />
+        <ProjectReference Include="..\DafnyCore\DafnyCore.csproj" />
     </ItemGroup>
 
     <ItemGroup>
-      <None Remove="ParsedAst" />
       <EmbeddedResource Include="ParsedAst">
         <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
       </EmbeddedResource>
